backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
alembic==1.13.1
psycopg2-binary==2.9.9
psycopg2[async]==2.9.9
asyncpg==0.29.0
python-dotenv==1.0.0
pydantic-settings==2.1.0
redis==5.0.0
celery==5.3.4
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
mypy==1.5.1

# AI Agent Execution Dependencies
httpx==0.25.2
yarl==1.9.2
pydantic-enum==0.8.0
certifi==2024.2.2
async-timeout==4.0.2
email-validator==2.1.0
python-multipart==0.0.6

# Slack Integration Dependencies
requests==2.31.0
certifi==2024.2.2

# Development Tools
flake8==6.1.0
black==23.11.0
isort==5.12.0
mypy==1.5.1
pre-commit==3.6.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
httpx-mock==0.11.1
pytest-cov==4.1.0
pytest-xdist==3.3.1